Fix search bug and return accurate search result count
There was an issue causing repeated searches (after the first search) to ignore the date range the user selected, and reverting back to the default start and end dates.
Also made the search result count returned independant of the number geoJSON objects built for the map, so that the result count is accurate, while limiting the total features rendered on the map.

Bugs:
- Repeat searches ignored user entered date range
- Search result count inaccurate and fixed to number of map features rendered